+++ src/compr_zstd.c
@@ -32,7 +32,7 @@ struct state {
 };
 
 static struct compr_state *
-compr_zstd_init(FILE *output, int level)
+compr_zstd_init(FILE *output, int8_t level)
 {
 	struct state *state = malloc(sizeof(struct state));
 	state->output = output;
